backend/requirements.txt
# backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic-core==2.10.1
motor==3.2.0
pymongo==4.12.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
email-validator==2.0.0.post2
httpx==0.25.1
python-dotenv==1.0.0
jinja2==3.1.2
dnspython-2.7.0